Modularized common use cases view
