+++ bridge/client/app/_components/_dialogs/ktb-delete-confirmation/ktb-delete-confirmation.module.ts
@@ -3,11 +3,10 @@ import { CommonModule } from '@angular/common';
 import { KtbDeleteConfirmationComponent } from './ktb-delete-confirmation.component';
 import { DtConfirmationDialogModule } from '@dynatrace/barista-components/confirmation-dialog';
 import { DtButtonModule } from '@dynatrace/barista-components/button';
-import { BrowserAnimationsModule } from '@angular/platform-browser/animations';
 
 @NgModule({
   declarations: [KtbDeleteConfirmationComponent],
-  imports: [CommonModule, BrowserAnimationsModule, DtConfirmationDialogModule, DtButtonModule],
+  imports: [CommonModule, DtConfirmationDialogModule, DtButtonModule],
   exports: [KtbDeleteConfirmationComponent],
 })
 export class KtbDeleteConfirmationModule {}
