+++ bridge/client/app/_components/ktb-expandable-tile/ktb-expandable-tile.module.ts
@@ -4,13 +4,11 @@ import { KtbExpandableTileComponent, KtbExpandableTileHeaderDirective } from './
 import { DtExpandablePanelModule } from '@dynatrace/barista-components/expandable-panel';
 import { DtShowMoreModule } from '@dynatrace/barista-components/show-more';
 import { DtIconModule } from '@dynatrace/barista-components/icon';
-import { BrowserAnimationsModule } from '@angular/platform-browser/animations';
 
 @NgModule({
   declarations: [KtbExpandableTileComponent, KtbExpandableTileHeaderDirective],
   imports: [
     CommonModule,
-    BrowserAnimationsModule,
     DtExpandablePanelModule,
     DtIconModule.forRoot({
       svgIconLocation: `assets/icons/{{name}}.svg`,
