--- bridge/client/app/_components/ktb-keptn-services-list/ktb-keptn-services-list.component.html
@@ -1,131 +1,0 @@
-<div fxFlexFill fxLayout="column" class="container" *ngIf="projectName as projectName">
-  <div fxLayout="column" fxFlex="0 0 60%" class="overflow-y-scroll">
-    <ktb-loading-distractor *ngIf="isLoadingUniformRegistrations">Loading ...</ktb-loading-distractor>
-
-    <div *ngIf="!isLoadingUniformRegistrations && uniformRegistrations.data && uniformRegistrations.data.length === 0">
-      No integrations available.
-    </div>
-
-    <dt-table
-      uitestid="keptn-uniform-integrations-table"
-      [dataSource]="uniformRegistrations"
-      dtSort
-      (dtSortChange)="sortData($event)"
-      interactiveRows
-      *ngIf="uniformRegistrations.data && uniformRegistrations.data.length > 0"
-    >
-      <ng-container dtColumnDef="name" dtColumnAlign="text" dtColumnProportion="2">
-        <dt-header-cell dt-sort-header *dtHeaderCellDef>Name</dt-header-cell>
-        <dt-cell *dtCellDef="let row">
-          <ng-container *ngIf="row | toType: UniformRegistrationClass as registration">
-            <span [textContent]="registration.name || 'n/a'"></span>
-            <div
-              *ngIf="registration.unreadEventsCount"
-              class="notification-indicator-text ml-2"
-              [textContent]="registration.unreadEventsCount"
-            ></div>
-          </ng-container>
-        </dt-cell>
-      </ng-container>
-
-      <ng-container dtColumnDef="version" dtColumnAlign="text" dtColumnProportion="2">
-        <dt-header-cell *dtHeaderCellDef>Version</dt-header-cell>
-        <dt-cell
-          *dtCellDef="let row"
-          [textContent]="(row | toType: UniformRegistrationClass).metadata.integrationversion || 'n/a'"
-        ></dt-cell>
-      </ng-container>
-
-      <ng-container dtColumnDef="host" dtColumnAlign="text" dtColumnProportion="3">
-        <dt-header-cell dt-sort-header *dtHeaderCellDef>Cluster / Host</dt-header-cell>
-        <dt-cell
-          *dtCellDef="let row"
-          [textContent]="(row | toType: UniformRegistrationClass).metadata.hostname || 'n/a'"
-        ></dt-cell>
-      </ng-container>
-
-      <ng-container dtColumnDef="namespace" dtColumnAlign="text" dtColumnProportion="2">
-        <dt-header-cell dt-sort-header *dtHeaderCellDef>Namespace</dt-header-cell>
-        <dt-cell
-          *dtCellDef="let row"
-          [textContent]="(row | toType: UniformRegistrationClass).metadata.kubernetesmetadata.namespace || 'n/a'"
-        ></dt-cell>
-      </ng-container>
-
-      <ng-container dtColumnDef="location" dtColumnAlign="text" dtColumnProportion="2">
-        <dt-header-cell dt-sort-header *dtHeaderCellDef>Location</dt-header-cell>
-        <dt-cell
-          *dtCellDef="let row"
-          [textContent]="(row | toType: UniformRegistrationClass).metadata.location || 'n/a'"
-        ></dt-cell>
-      </ng-container>
-
-      <ng-container dtColumnDef="subscription" dtColumnAlign="text" dtColumnProportion="3">
-        <dt-header-cell *dtHeaderCellDef>Active subscriptions</dt-header-cell>
-        <dt-cell
-          *dtCellDef="let row"
-          [innerHtml]="formatSubscriptions(row | toType: UniformRegistrationClass, projectName) || 'n/a'"
-        ></dt-cell>
-      </ng-container>
-
-      <dt-header-row
-        *dtHeaderRowDef="['name', 'version', 'host', 'namespace', 'location', 'subscription']"
-      ></dt-header-row>
-      <dt-row
-        *dtRowDef="let row; columns: ['name', 'version', 'host', 'namespace', 'location', 'subscription']"
-        class="pointer"
-        [class.active]="selectedUniformRegistration === row"
-        [attr.disabled]="!(row | toType: UniformRegistrationClass).hasSubscriptions(projectName)"
-        [dtOverlay]="getOverlay(row, projectName, noRegistration)"
-        (click)="setSelectedUniformRegistration(row)"
-      >
-        <ng-template #noRegistration>
-          <span>{{ (row | toType: UniformRegistrationClass).name }} is not active for this project.</span>
-          <br />
-          <span>Add subscription to activate it for {{ projectName }}.</span>
-        </ng-template>
-      </dt-row>
-    </dt-table>
-  </div>
-  <div fxLayout="column" fxFlex="0 0 40%">
-    <div class="no-integration-selected p-3 mt-3" fxLayout="row" *ngIf="!selectedUniformRegistration">
-      <dt-icon class="event-icon info mr-1" name="information"></dt-icon>
-      Select an integration to show error events
-    </div>
-
-    <ng-container *ngIf="selectedUniformRegistration">
-      <h3 class="mt-3 mb-3"><span class="mr-1 bold" [textContent]="selectedUniformRegistration.name"></span></h3>
-
-      <ktb-expandable-tile
-        [alignment]="'left'"
-        class="mb-2"
-        [expanded]="true"
-        *ngIf="selectedUniformRegistration.isChangeable()"
-      >
-        <ktb-expandable-tile-header>
-          <h3 class="mt-0 mb-0">Subscriptions</h3>
-        </ktb-expandable-tile-header>
-        <ktb-uniform-subscriptions [uniformRegistration]="selectedUniformRegistration"></ktb-uniform-subscriptions>
-      </ktb-expandable-tile>
-
-      <div class="pb-4">
-        <ktb-expandable-tile [alignment]="'left'" [expanded]="true">
-          <ktb-expandable-tile-header>
-            <h3 class="mt-0 mb-0">Error events</h3>
-          </ktb-expandable-tile-header>
-
-          <ktb-loading-distractor *ngIf="isLoadingLogs">Loading ...</ktb-loading-distractor>
-
-          <div fxLayout="column" *ngIf="!isLoadingLogs">
-            <ktb-uniform-registration-logs
-              *ngIf="uniformRegistrationLogs$ | async as logs"
-              [projectName]="projectName"
-              [lastSeen]="lastSeen"
-              [logs]="logs"
-            ></ktb-uniform-registration-logs>
-          </div>
-        </ktb-expandable-tile>
-      </div>
-    </ng-container>
-  </div>
-</div>
