--- bridge/client/app/_components/ktb-markdown/ktb-markdown.component.scss
@@ -1,2 +1,0 @@
-@import '../../../variables';
-@import '~@dynatrace/barista-components/core/src/style/variables';
