+++ bridge/client/app/_views/ktb-common-use-cases-view/ktb-common-use-cases-view.component.html
@@ -1,0 +1,76 @@
+<div fxFlexFill>
+  <div id="integration-container" class="container" fxFlex fxLayout="column">
+    <p>
+      Integrate Keptn using the <a [href]="'/reference/cli/' | keptnUrl">Keptn CLI</a> or
+      <a [href]="'/reference/api/' | keptnUrl">Keptn API</a>.
+    </p>
+    <h2>Keptn CLI</h2>
+    <ol>
+      <li>
+        <a
+          *ngIf="keptnInfo?.bridgeInfo?.cliDownloadLink"
+          [attr.href]="keptnInfo?.bridgeInfo?.cliDownloadLink"
+          rel="noopener noreferrer"
+          target="_blank"
+          >Download Keptn CLI</a
+        >
+      </li>
+      <li>
+        Authenticate CLI with:
+        <a [href]="'/reference/cli/commands/keptn_auth/' | keptnUrl" rel="noopener noreferrer" target="_blank"
+          >keptn auth</a
+        >
+      </li>
+    </ol>
+    <p class="bold mt-2" *ngIf="useCaseExamples.cli.length > 0">Use Cases:</p>
+    <ul class="list-style-none">
+      <li *ngFor="let example of useCaseExamples.cli">
+        <a href="javascript:void(0);" (click)="exampleCode.toggle()"
+          ><button class="inline small mr-2" dt-show-more [showLess]="exampleCode.expanded"></button>
+          <span [textContent]="example.label"></span
+        ></a>
+        <dt-expandable-panel #exampleCode>
+          <pre class="code"><code class="lang-console" [innerHtml]="example.code"></code></pre>
+        </dt-expandable-panel>
+      </li>
+    </ul>
+    <h2>Keptn API</h2>
+    <ol>
+      <ng-container *ngIf="keptnInfo && keptnInfo.bridgeInfo && keptnInfo.bridgeInfo.showApiToken; else noApiToken">
+        <li><a href="javascript:void(0);" (click)="copyApiToken()">Copy API token to clipboard</a></li>
+        <li>Access <a [href]="keptnInfo.bridgeInfo.apiUrl" rel="noopener noreferrer" target="_blank">Keptn API</a></li>
+      </ng-container>
+      <ng-template #noApiToken>
+        <li>
+          <a [href]="'/operate/api_token/' | keptnUrl" rel="noopener noreferrer" target="_blank">Retrieve API token</a>
+        </li>
+        <li>Access Keptn API</li>
+      </ng-template>
+    </ol>
+    <p class="bold mt-2" *ngIf="useCaseExamples.api.length > 0">Use Cases:</p>
+    <ul class="list-style-none">
+      <li *ngFor="let example of useCaseExamples.api">
+        <a href="javascript:void(0);" (click)="exampleCode.toggle()"
+          ><button class="inline small mr-2" dt-show-more [showLess]="exampleCode.expanded"></button>
+          <span [textContent]="example.label"></span
+        ></a>
+        <dt-expandable-panel #exampleCode>
+          <pre class="code"><code class="lang-console" [innerHtml]="example.code"></code></pre>
+        </dt-expandable-panel>
+      </li>
+    </ul>
+    <p *ngIf="!integrationsExternalDetails">
+      <button dt-button (click)="loadIntegrations()">
+        <dt-icon name="xhr-load-action"></dt-icon> Load more information about integrations
+      </button>
+    </p>
+    <ktb-markdown [html]="integrationsExternalDetails"></ktb-markdown>
+    <p class="small light">
+      Additional information about integrations will be loaded from external sources and Keptn will collect statistical
+      data. Details can be found at
+      <a [href]="'/reference/load_information/' | keptnUrl" target="_blank" rel="noopener noreferrer">{{
+        '/reference/load_information/' | keptnUrl
+      }}</a>
+    </p>
+  </div>
+</div>
