+++ bridge/client/app/_views/ktb-common-use-cases-view/ktb-common-use-cases-view.component.scss
@@ -1,0 +1,11 @@
+#integration-container {
+  ::ng-deep p,
+  ul,
+  ol {
+    margin: 0;
+  }
+  h2 {
+    margin-top: 0.3em;
+    margin-bottom: 0.3em;
+  }
+}
