+++ bridge/client/app/_views/ktb-common-use-cases-view/ktb-markdown/ktb-markdown.component.html
