+++ bridge/client/app/_views/ktb-common-use-cases-view/ktb-markdown/ktb-markdown.component.scss
@@ -1,0 +1,2 @@
+@import 'client/variables';
+@import 'node_modules/@dynatrace/barista-components/core/src/style/variables';
