+++ bridge/client/app/_views/ktb-integration-view/ktb-integration-view.component.spec.ts
@@ -1,19 +1,37 @@
 import { ComponentFixture, TestBed } from '@angular/core/testing';
 import { KtbIntegrationViewComponent } from './ktb-integration-view.component';
 import { HttpClientTestingModule } from '@angular/common/http/testing';
-import { POLLING_INTERVAL_MILLIS } from '../../_utils/app.utils';
-import { AppModule } from '../../app.module';
+import { ActivatedRoute, convertToParamMap } from '@angular/router';
+import { of } from 'rxjs';
+import { ApiService } from '../../_services/api.service';
+import { ApiServiceMock } from '../../_services/api.service.mock';
+import { RouterTestingModule } from '@angular/router/testing';
+import { KtbCommonUseCasesViewModule } from '../ktb-common-use-cases-view/ktb-common-use-cases-view.module';
 
-describe('KtbIntegrationViewComponent', () => {
+describe('KtbKeptnServicesListComponent', () => {
   let component: KtbIntegrationViewComponent;
   let fixture: ComponentFixture<KtbIntegrationViewComponent>;
 
   beforeEach(async () => {
     await TestBed.configureTestingModule({
-      imports: [AppModule, HttpClientTestingModule],
-      providers: [{ provide: POLLING_INTERVAL_MILLIS, useValue: 0 }],
+      declarations: [],
+      imports: [KtbCommonUseCasesViewModule, HttpClientTestingModule, RouterTestingModule],
+      providers: [
+        { provide: ApiService, useClass: ApiServiceMock },
+        {
+          provide: ActivatedRoute,
+          useValue: {
+            paramMap: of(
+              convertToParamMap({
+                projectName: 'sockshop',
+              })
+            ),
+          },
+        },
+      ],
     }).compileComponents();
 
+    localStorage.setItem('keptn_integration_dates', '');
     fixture = TestBed.createComponent(KtbIntegrationViewComponent);
     component = fixture.componentInstance;
     fixture.detectChanges();
