+++ bridge/client/app/_views/ktb-integration-view/ktb-integration-view.module.ts
@@ -1,0 +1,50 @@
+import { NgModule } from '@angular/core';
+import { KtbIntegrationViewRoutingModule } from './ktb-integration-view-routing.module';
+import { ReactiveFormsModule } from '@angular/forms';
+import { FlexLayoutModule } from '@angular/flex-layout';
+import { KtbLoadingModule } from '../../_components/ktb-loading/ktb-loading.module';
+import { KtbWebhookSettingsModule } from '../../_components/ktb-webhook-settings/ktb-webhook-settings.module';
+import { KtbPayloadViewerModule } from '../../_components/ktb-payload-viewer/ktb-payload-viewer.module';
+import { DtSelectModule } from '@dynatrace/barista-components/select';
+import { DtOverlayModule } from '@dynatrace/barista-components/overlay';
+import { DtFilterFieldModule } from '@dynatrace/barista-components/filter-field';
+import { DtFormFieldModule } from '@dynatrace/barista-components/form-field';
+import { DtIconModule } from '@dynatrace/barista-components/icon';
+import { DtCheckboxModule } from '@dynatrace/barista-components/checkbox';
+import { DtButtonModule } from '@dynatrace/barista-components/button';
+import { CommonModule } from '@angular/common';
+import { DtTableModule } from '@dynatrace/barista-components/table';
+import { KtbExpandableTileModule } from '../../_components/ktb-expandable-tile/ktb-expandable-tile.module';
+import { KtbPipeModule } from '../../_pipes/ktb-pipe.module';
+import { KtbUniformRegistrationLogsModule } from '../../_components/ktb-uniform-registration-logs/ktb-uniform-registration-logs.module';
+import { KtbUniformSubscriptionsModule } from '../../_components/ktb-uniform-subscriptions/ktb-uniform-subscriptions.module';
+import { KtbIntegrationViewComponent } from './ktb-integration-view.component';
+import { KtbModifyUniformSubscriptionComponent } from './ktb-modify-uniform-subscription/ktb-modify-uniform-subscription.component';
+
+@NgModule({
+  declarations: [KtbIntegrationViewComponent, KtbModifyUniformSubscriptionComponent],
+  imports: [
+    CommonModule,
+    DtButtonModule,
+    DtCheckboxModule,
+    DtIconModule.forRoot({
+      svgIconLocation: `assets/icons/{{name}}.svg`,
+    }),
+    DtFormFieldModule,
+    DtFilterFieldModule,
+    DtOverlayModule,
+    DtSelectModule,
+    KtbPayloadViewerModule,
+    KtbWebhookSettingsModule,
+    KtbLoadingModule,
+    FlexLayoutModule,
+    ReactiveFormsModule,
+    KtbIntegrationViewRoutingModule,
+    DtTableModule,
+    KtbExpandableTileModule,
+    KtbPipeModule,
+    KtbUniformRegistrationLogsModule,
+    KtbUniformSubscriptionsModule,
+  ],
+})
+export class KtbIntegrationViewModule {}
