+++ bridge/client/app/_views/ktb-integration-view/ktb-modify-uniform-subscription/ktb-modify-uniform-subscription.component.spec.ts
@@ -1,22 +1,22 @@
 import { ComponentFixture, TestBed } from '@angular/core/testing';
 import { KtbModifyUniformSubscriptionComponent } from './ktb-modify-uniform-subscription.component';
 import { ActivatedRoute, convertToParamMap, ParamMap, Router } from '@angular/router';
-import { UniformRegistrationsMock } from '../../_services/_mockData/uniform-registrations.mock';
+import { UniformRegistrationsMock } from '../../../_services/_mockData/uniform-registrations.mock';
 import { BehaviorSubject, of, throwError } from 'rxjs';
-import { DataService } from '../../_services/data.service';
-import { UniformSubscription } from '../../_models/uniform-subscription';
+import { DataService } from '../../../_services/data.service';
+import { UniformSubscription } from '../../../_models/uniform-subscription';
 import { HttpClientTestingModule } from '@angular/common/http/testing';
-import { UniformRegistrationLocations } from '../../../../shared/interfaces/uniform-registration-locations';
-import { UniformRegistrationInfo } from '../../../../shared/interfaces/uniform-registration-info';
+import { UniformRegistrationLocations } from '../../../../../shared/interfaces/uniform-registration-locations';
+import { UniformRegistrationInfo } from '../../../../../shared/interfaces/uniform-registration-info';
 import { HttpErrorResponse } from '@angular/common/http';
 import { AbstractControl } from '@angular/forms';
-import { ApiService } from '../../_services/api.service';
-import { ApiServiceMock } from '../../_services/api.service.mock';
-import { ProjectsMock } from '../../_services/_mockData/projects.mock';
-import { KtbModifyUniformSubscriptionModule } from './ktb-modify-uniform-subscription.module';
+import { ApiService } from '../../../_services/api.service';
+import { ApiServiceMock } from '../../../_services/api.service.mock';
+import { ProjectsMock } from '../../../_services/_mockData/projects.mock';
 import { RouterTestingModule } from '@angular/router/testing';
-import { KtbKeptnServicesListComponent } from '../ktb-keptn-services-list/ktb-keptn-services-list.component';
-import { IWebhookConfigClient } from '../../../../shared/interfaces/webhook-config';
+import { KtbIntegrationViewComponent } from '../ktb-integration-view.component';
+import { IWebhookConfigClient } from '../../../../../shared/interfaces/webhook-config';
+import { KtbCommonUseCasesViewModule } from '../../ktb-common-use-cases-view/ktb-common-use-cases-view.module';
 
 describe('KtbModifyUniformSubscriptionComponent', () => {
   let component: KtbModifyUniformSubscriptionComponent;
@@ -32,12 +32,12 @@ describe('KtbModifyUniformSubscriptionComponent', () => {
     );
     await TestBed.configureTestingModule({
       imports: [
-        KtbModifyUniformSubscriptionModule,
+        KtbCommonUseCasesViewModule,
         HttpClientTestingModule,
         RouterTestingModule.withRoutes([
           {
             path: 'project/:projectName/settings/uniform/integrations/:integrationId',
-            component: KtbKeptnServicesListComponent,
+            component: KtbIntegrationViewComponent,
           },
         ]),
       ],
